requirements.txt
annotated-types     ~= 0.6.0
anyio               ~=  4.3.0
click               ~=  8.1.7
fastapi             ~=  0.110.0
greenlet            ~=  3.0.3
h11                 ~=  0.14.0
idna                ~=  3.6
pip                 ~=  24.0
pydantic            ~=  2.6.3
pydantic_core       ~=  2.16.3
setuptools          ~=  65.5.0
sniffio              ~=  1.3.1
SQLAlchemy          ~=  2.0.27
starlette           ~=  0.36.3
typing_extensions   ~= 4.10.0
uvicorn             ~= 0.27.1
python-jose         ~= 3.3.0
passlib             ~= 1.7.4
python-multipart    ~=0.0.9
bcrypt              ~=4.1.2
google-cloud-storage ~=  2.16.0
python-decouple      ~=    3.8